moved DTO classes to DTO folder
