+++ sdm-iso-back/src/main/java/com/example/sdmisoback/dto/AttachmentFileView.java
@@ -1,4 +1,4 @@
-package com.example.sdmisoback.view;
+package com.example.sdmisoback.dto;
 
 import java.time.LocalDateTime;
