+++ sdm-iso-back/src/main/java/com/example/sdmisoback/dto/FiltersDTO.java
@@ -1,4 +1,4 @@
-package com.example.sdmisoback.model;
+package com.example.sdmisoback.dto;
 
 import java.time.LocalDateTime;
