+++ sdm-iso-back/src/main/java/com/example/sdmisoback/repository/CustomAttachmentFileRepo.java
@@ -2,8 +2,8 @@ package com.example.sdmisoback.repository;
 
 import org.springframework.data.domain.Page;
 
-import com.example.sdmisoback.model.FiltersDTO;
-import com.example.sdmisoback.view.AttachmentFileView;
+import com.example.sdmisoback.dto.AttachmentFileView;
+import com.example.sdmisoback.dto.FiltersDTO;
 
 public interface CustomAttachmentFileRepo {
     Page<AttachmentFileView> filterAttachments(FiltersDTO filters);
